move to githubservice class
