+++ openhands/server/services/github_service.py
@@ -1,7 +1,15 @@
+import httpx
 import requests
+from fastapi import HTTPException
+from fastapi.responses import JSONResponse
+
+from openhands.server.shared import server_config
+from openhands.utils.async_utils import call_sync_from_async
 
 
 class GitHubService:
+    BASE_URL = 'https://api.github.com'
+
     def __init__(self, token: str):
         self.token = token
         self.headers = {
@@ -9,8 +17,79 @@ class GitHubService:
             'Accept': 'application/vnd.github.v3+json',
         }
 
-    def get_user(self):
-        response = requests.get('https://api.github.com/user', headers=self.headers)
-        response.raise_for_status()
+    # def get_user(self):
+    #     response = requests.get('https://api.github.com/user', headers=self.headers)
+    #     response.raise_for_status()
 
-        return response.json()
+    #     return response.json()
+
+    def _should_refresh(self, status_code: int):
+        if server_config.app_mode == 'SAAS' and status_code == 401:
+            return True
+
+        return False
+
+    async def _refresh_token(self):
+        pass
+
+    async def _fetch_data(self, url: str, params: dict | None = None):
+        try:
+            async with httpx.AsyncClient() as client:
+                response = await client.get(url, headers=self.headers, params=params)
+                if self._should_refresh(response.status_code):
+                    await self._refresh_token()
+                    response = await client.get(
+                        url, headers=self.headers, params=params
+                    )
+                response.raise_for_status()
+                return response
+
+        except httpx.HTTPStatusError as e:
+            status_code = e.response.status_code
+            error_detail = e.response.text
+
+            raise HTTPException(
+                status_code=status_code,
+                detail=f'GitHub API error: {error_detail}',
+            )
+        except httpx.HTTPError as e:
+            raise HTTPException(status_code=500, detail=f'HTTP error: {str(e)}')
+
+    async def get_user(self):
+        url = f'{self.BASE_URL}/user'
+        return await self._fetch_data(url)
+
+    async def get_repositories(
+        self, page: int, per_page: int, sort: str, installation_id: int | None
+    ):
+        params = {'page': str(page), 'per_page': str(per_page)}
+        if installation_id:
+            url = f'{self.BASE_URL}/user/installations/{installation_id}/repositories'
+        else:
+            url = f'{self.BASE_URL}/user/repos'
+            params['sort'] = sort
+        return await self._fetch_data(url, params)
+
+    async def get_installation_ids(self):
+        url = f'{self.BASE_URL}/user/installations'
+        response = await self._fetch_data(url)
+        data = response.json()
+        return data.get('installations', [])
+
+    async def search_repositories(
+        self, query: str, per_page: int, sort: str, order: str
+    ):
+        url = f'{self.BASE_URL}/search/repositories'
+        params = {'q': query, 'per_page': per_page, 'sort': sort, 'order': order}
+        return await call_sync_from_async(
+            requests.get, url, headers=self.headers, params=params
+        )
+
+    async def fetch_response(self, method: str, *args, **kwargs):
+        response = await getattr(self, method)(*args, **kwargs)
+        json_response = JSONResponse(
+            content=response.json(), status_code=response.status_code
+        )
+        if 'Link' in response.headers:
+            json_response.headers['Link'] = response.headers['Link']
+        return json_response
